+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 1. [11678](https://github.com/influxdata/influxdb/pull/11678): Update the System Telegraf Plugin bundle to include the swap plugin
 
 ## UI Improvements
-
+1. [11683](https://github.com/influxdata/influxdb/pull/11683): Change the wording for the plugin config form button to Done
 
 ## v2.0.0-alpha.1 [2019-01-23]